+++ resources/browse/category.php
@@ -100,6 +100,10 @@
                 </a>
                 <div class="caption">
                     <a href="/image/<?= h($image['id']) ?>"><?= h($image['descr']) ?></a>
+                    <div class="caption-meta">
+                        <small class="caption-left"><?= $photographer ? h($photographer) : '' ?></small>
+                        <small class="caption-right"><?= $image['date_taken'] ? format_date($image['date_taken']) : '' ?></small>
+                    </div>
                 </div>
             </div>
         <?php endforeach; ?>